Make NixOS module work, secrets handling, security hardening
We want to generate a config and put it in the Nix store but we
don't want to expose secrets there.

Secrets are the oauth client secret and the notify secret at the moment.

Uses various security hardening features for SystemD units.
Score in systemd-analyze security is already OK, but could be improved
with more syscall filtering and a chroot, for example.